thermal: netlink: Fix compilation error when CONFIG_NET=n
When the network is not configured, the netlink is disabled on all
the system. The thermal framework assumed the netlink is always
opt-in.

Fix this by adding a Kconfig option for the netlink notification,
defaulting to yes and depending on CONFIG_NET.

As the change implies multiple stubs and in order to not pollute the
internal thermal header, the thermal_nelink.h has been added and
included in the thermal_core.h, so this one regain some kind of
clarity.

Reported-by: Randy Dunlap <rdunlap@infradead.org>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Reviewed-by: Amit Kucheria <amit.kucheria@linaro.org>
Link: https://lore.kernel.org/r/20200707090159.1018-1-daniel.lezcano@linaro.org

(cherry picked from commit 5b8583d3bd7fc10cea07e4a5bfa59465758a39dc)
Signed-off-by: Madara273 <ravenhoxs@gmail.com>
(cherry picked from commit 13f2972c0b87f7ab50a6154565d9dbf8840fba89)
Signed-off-by: Madara273 <ravenhoxs@gmail.com>