torture: Fix kvm.sh --datestamp regex check
Some versions of grep are happy to interpret a nonsensically placed "-"
within a "[]" pattern as a dash, while others give an error message.
This commit therefore places the "-" at the end of the expression where
it was supposed to be in the first place.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
(cherry picked from commit 860d25f792fb78d2379e0be36c495bc9631945ec)
Signed-off-by: Madara273 <ravenhoxs@gmail.com>